refactor(media): split transcription metrics responsibility
