+++ backend-spring/src/main/java/com/myway/backendspring/feature/media/MediaTranscriptionService.java
@@ -22,7 +22,6 @@ public class MediaTranscriptionService {
     private static final String STT_DEFAULT_MODEL = "cf-whisper";
     private static final int PUBLIC_STT_MAX_DURATION_MS = 180_000;
     private static final int FALLBACK_TRANSCRIPT_DURATION_MS = 120_000;
-    private static final int STT_TARGET_SEGMENT_WORDS = 20;
     private static final String SPEAKER_REVIEW_SCOPE = "media_speaker_review";
 
     private final FeatureStoreRepository repository;
@@ -30,19 +29,22 @@ public class MediaTranscriptionService {
     private final ActivityEventService activityEventService;
     private final RagService ragService;
     private final TranscriptSegmenter segmenter;
+    private final MediaTranscriptionMetrics metrics;
 
     public MediaTranscriptionService(
             FeatureStoreRepository repository,
             DemoLearningService learningService,
             ActivityEventService activityEventService,
             RagService ragService,
-            TranscriptSegmenter segmenter
+            TranscriptSegmenter segmenter,
+            MediaTranscriptionMetrics metrics
     ) {
         this.repository = repository;
         this.learningService = learningService;
         this.activityEventService = activityEventService;
         this.ragService = ragService;
         this.segmenter = segmenter;
+        this.metrics = metrics;
     }
 
     public Map<String, Object> transcribe(
@@ -141,8 +143,8 @@ public class MediaTranscriptionService {
         Map<String, Object> instructorGuess = inferInstructorGuess(lectureId);
         List<Map<String, Object>> speakerSegments = buildSpeakerSegments(segments, instructorGuess);
         String transcriptId = String.valueOf(extraction.getOrDefault("id", UUID.randomUUID().toString()));
-        int wordCount = countWords(text);
-        Map<String, Object> quality = sttQualityMetrics(segments, request.durationMs());
+        int wordCount = metrics.countWords(text);
+        Map<String, Object> quality = metrics.sttQualityMetrics(segments, request.durationMs());
         return new TranscriptionDraft(
                 transcriptId,
                 lectureId,
@@ -456,48 +458,6 @@ public class MediaTranscriptionService {
                 "특히 자주 혼동되는 개념을 예시로 비교하고, 다음 차시와 연결되는 질문을 중심으로 학습합니다.";
     }
 
-    private Map<String, Object> sttQualityMetrics(List<Map<String, Object>> segments, int durationMs) {
-        if (segments == null || segments.isEmpty()) {
-            return Map.of("segment_count", 0, "avg_words_per_segment", 0, "min_segment_ms", 0, "max_segment_ms", 0, "target_segment_words", STT_TARGET_SEGMENT_WORDS, "quality_score", 0.0);
-        }
-        int totalWords = 0;
-        int minMs = Integer.MAX_VALUE;
-        int maxMs = 0;
-        for (Map<String, Object> segment : segments) {
-            int start = asInt(segment.get("start_ms"));
-            int end = asInt(segment.get("end_ms"));
-            int ms = Math.max(0, end - start);
-            minMs = Math.min(minMs, ms);
-            maxMs = Math.max(maxMs, ms);
-            totalWords += countWords(String.valueOf(segment.getOrDefault("text", "")));
-        }
-        double avgWords = totalWords / (double) Math.max(1, segments.size());
-        double wordFit = 1.0 - Math.min(1.0, Math.abs(avgWords - STT_TARGET_SEGMENT_WORDS) / STT_TARGET_SEGMENT_WORDS);
-        double durationFit = 1.0 - Math.min(1.0, Math.abs(durationMs - (segments.size() * 20_000.0)) / Math.max(1.0, durationMs));
-        double score = Math.max(0.0, Math.min(1.0, (wordFit * 0.7) + (durationFit * 0.3)));
-        return Map.of(
-                "segment_count", segments.size(),
-                "avg_words_per_segment", Math.round(avgWords * 100.0) / 100.0,
-                "min_segment_ms", minMs == Integer.MAX_VALUE ? 0 : minMs,
-                "max_segment_ms", maxMs,
-                "target_segment_words", STT_TARGET_SEGMENT_WORDS,
-                "quality_score", Math.round(score * 1000.0) / 1000.0
-        );
-    }
-
-    private int countWords(String text) {
-        return tokenize(text).size();
-    }
-
-    private List<String> tokenize(String text) {
-        if (text == null) return List.of();
-        String normalized = text.replaceAll("\\s+", " ").trim().toLowerCase();
-        return List.of(normalized.split("[^a-zA-Z0-9가-힣]+")).stream()
-                .map(String::trim)
-                .filter(token -> token.length() > 1)
-                .toList();
-    }
-
     private void putIfText(Map<String, Object> target, String key, String value) {
         if (value != null && !value.isBlank()) target.put(key, value);
     }
@@ -506,16 +466,6 @@ public class MediaTranscriptionService {
         if (value == null) return defaultValue;
         String trimmed = value.trim();
         return trimmed.isEmpty() ? defaultValue : trimmed;
-    }
-
-    private int asInt(Object value) {
-        if (value == null) return 0;
-        if (value instanceof Number n) return n.intValue();
-        try {
-            return Integer.parseInt(String.valueOf(value));
-        } catch (Exception ignored) {
-            return 0;
-        }
     }
 
     private long asLong(Object value) {
